bench/web/requirements.txt
duckdb>=1.0
psycopg2-binary
clickhouse-connect
influxdb-client
influxdb3-python
pyarrow
pandas
fastapi
uvicorn[standard]
requests
questdb
pydantic-settings
psutil
